Change the way we store method definitions to disambiguate aliases
Aliased methods seem to have equal `Method` objects associated with them, so we can't use the method object the a key in a hash. Instead, we should store the method definitions in a hash keyed by the owner of the method, and then the method name. This way, we can disambiguate between aliased methods.